@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "organize-tool"
-version = "2.4.3"
+version = "3.0.0a0"
 description = "The file management automation tool"
 packages = [{ include = "organize" }]
 authors = ["Thomas Feldmann <mail@tfeldmann.de>"]
@@ -24,6 +24,7 @@ classifiers = [
     "Intended Audience :: End Users/Desktop",
     "Natural Language :: English",
     "Operating System :: OS Independent",
+    "Programming Language :: Python :: 3",
     "Topic :: Utilities",
 ]
 
@@ -31,25 +32,26 @@ classifiers = [
 organize = "organize.cli:cli"
 
 [tool.poetry.dependencies]
-python = ">=3.8,<3.13"
-fs = ">=2.4.16"
-rich = "^13.4.2"
-PyYAML = "^6.0"
-Send2Trash = { version = "^1.8.2", extras = ["nativeLib"] }
-textract = { version = "^1.6.5", optional = true }
-simplematch = "^1.3"
-macos-tags = { version = "^1.5.1", markers = "sys_platform == 'darwin'" }
-schema = "^0.7.5"
+python = "^3.9"
+arrow = "^1.3.0"
+docopt-ng = "^0.9.0"
+ExifRead = "2.3.2"                                                        # Pinned: https://github.com/tfeldmann/organize/issues/267
 Jinja2 = "^3.1.2"
-click = "^8.1.3"
-ExifRead = "^2.3.2" # Pinned: https://github.com/tfeldmann/organize/issues/267
+macos-tags = { version = "^1.5.1", markers = "sys_platform == 'darwin'" }
+platformdirs = "^4.0.0"
+pydantic = "^2.3.0"
+PyYAML = "^6.0"
+rich = "^13.4.2"
+Send2Trash = { version = "^1.8.2", extras = ["nativeLib"] }
+simplematch = "^1.4"
+textract = { version = "^1.6.5", optional = true }
 
 # must be in main dependencies for readthedocs.
-mkdocs = { version = "^1.2.3", optional = true }
+mkdocs = { version = "^1.5.3", optional = true }
 mkdocs-autorefs = { version = "^0.5.0", optional = true }
 mkdocs-include-markdown-plugin = { version = "^6.0.0", optional = true }
-mkdocstrings = { version = "^0.22.0", extras = ["python"], optional = true }
-markupsafe = { version = "2.0.1", optional = true } # Pinned: https://stackoverflow.com/q/72191560/300783
+mkdocstrings = { version = "^0.24.0", extras = ["python"], optional = true }
+markupsafe = { version = "2.0.1", optional = true }                          # Pinned: https://stackoverflow.com/q/72191560/300783
 
 [tool.poetry.extras]
 textract = ["textract"]
@@ -68,13 +70,18 @@ requests = "^2.31.0"
 types-PyYAML = "^6.0.12.10"
 coverage = "^7.2.0"
 pytest-cov = "^4.1.0"
-ruff = ">=0.0.285,<0.0.292"
+ruff = "^0.1.5"
+pyfakefs = "^5.3.1"
 
 [tool.coverage.run]
 source = ['organize']
 
+[tool.coverage.report]
+exclude_also = ["pragma: no cover", "if TYPE_CHECKING:"]
+
 [tool.mypy]
-python_version = "3.8"
+python_version = "3.9"
+plugins = ["pydantic.mypy"]
 
 [[tool.mypy.overrides]]
 module = [
@@ -93,6 +100,7 @@ ignore_missing_imports = true
 addopts = "--doctest-modules"
 testpaths = ["tests", "organize"]
 norecursedirs = ["tests/todo", "organize/filters", ".configs"]
+filterwarnings = ["ignore::DeprecationWarning"]
 
 [tool.isort]
 profile = "black"
