@@ -1,360 +1,253 @@
-"""
-organize
+__doc__ = """
+organize - The file management automation tool.
 
-The file management automation tool.
+Usage:
+  organize run   [options] [<config>]
+  organize sim   [options] [<config>]
+  organize new   [<config>]
+  organize edit  [<config>]
+  organize check [<config>]
+  organize debug [<config>]
+  organize show  [--path|--reveal] [<config>]
+  organize list
+  organize docs
+  organize --version
+  organize --help
+
+Commands:
+  run        Organize your files.
+  sim        Simulate organizing your files.
+  new        Creates a default config.
+  edit       Edit the config file with $EDITOR
+  check      Check config file validity
+  debug      Shows the raw config parsing steps.
+  show       Print the config to stdout.
+               Use --reveal to reveal the file in your file manager
+               Use --path to show the path to the file
+  list       Lists config files found in the default locations.
+  docs       Open the documentation.
+
+Options:
+  <config>                        A config name or path to a config file
+  -W --working-dir <dir>          The working directory
+  -F --format (DEFAULT|JSONL)     The output format [Default: DEFAULT]
+  -T --tags <tags>                Tags to run (eg. "initial,release")
+  -S --skip-tags <tags>           Tags to skip
+  -h --help                       Show this help page.
 """
 import os
 import sys
-import textwrap
-from typing import Optional, Tuple
+from functools import partial
+from pathlib import Path
+from typing import Literal, Optional, Set
 
-import click
-import fs
+from docopt import docopt
+from pydantic import BaseModel, ConfigDict, Field, ValidationError, field_validator
+from pydantic.functional_validators import BeforeValidator
+from rich.console import Console
+from rich.syntax import Syntax
+from rich.table import Table
+from typing_extensions import Annotated
+from yaml.scanner import ScannerError
 
-from . import console
-from .__version__ import __version__
-from .migration import NeedsMigrationError
+from organize import Config, ConfigError
+from organize.find_config import ConfigNotFound, find_config, list_configs
+from organize.output import JSONL, Default
+
+from .__version__ import __is_prerelease__, __version__
 
 DOCS_RTD = "https://organize.readthedocs.io"
 DOCS_GHPAGES = "https://tfeldmann.github.io/organize/"
 
-DEFAULT_CONFIG_FS_URL = "userconf://organize::/config.yaml"
+EXAMPLE_CONFIG = f"""\
+# organize configuration file
+# {DOCS_RTD}
+
+rules:
+  - locations:
+    filters:
+    actions:
+      - echo: "Hello, World!"
+"""
+
+Tags = Set[str]
+OutputFormat = Annotated[
+    Literal["default", "jsonl"], BeforeValidator(lambda v: v.lower())
+]
+
+console = Console()
 
 
-def path_split(path_or_url: str) -> Tuple[str, str]:
-    if sys.platform.startswith("win"):
-        path_or_url = path_or_url.replace("\\", "/")
-    dirname, filename = fs.path.split(path_or_url)
-    return dirname, filename
+def _open_uri(uri: str):
+    import webbrowser
 
-
-def ensure_default_config():
-    """
-    Ensures a configuration file exists in the default location.
-    """
-    DEFAULT_CONFIG_TEXT = textwrap.dedent(
-        """\
-        # organize configuration file
-        # {docs}
-
-        rules:
-          - name: "The name of this rule"
-            locations:
-              - # your locations here
-            filters:
-              - # your filters here
-            actions:
-              - # your actions here
-        """
-    ).format(docs=DOCS_RTD)
-
-    dirname, filename = path_split(DEFAULT_CONFIG_FS_URL)
-    if not filename:
-        raise ValueError("invalid config path, missing filename")
-    with fs.open_fs(dirname, create=True, writeable=True) as confdir:
-        if not confdir.exists(filename):
-            confdir.writetext(filename, DEFAULT_CONFIG_TEXT)
-
-
-def config_path(fs_url: Optional[str] = None) -> Tuple[bool, str]:
-    """
-    Return the config path, resolved into a syspath if possible.
-    If no fs_url is given, the default locations are checked.
-    As last resort, a config is created in the default location.
-    """
-    is_syspath = False
-    if not fs_url:
-        # first check whether the user set a env var
-        env_fs_url = os.getenv("ORGANIZE_CONFIG")
-        if env_fs_url:
-            fs_url = env_fs_url
-        else:
-            # if no env variable is given we make sure that there is a config file in
-            # the default location
-            ensure_default_config()
-            fs_url = DEFAULT_CONFIG_FS_URL
-
-    dirname, filename = path_split(fs_url)
-    try:
-        with fs.open_fs(dirname) as confdir:
-            config_path = confdir.getsyspath(filename)
-            is_syspath = True
-    except Exception:
-        config_path = fs_url
-    return is_syspath, config_path
-
-
-def read_config(fs_url: Optional[str] = None) -> Tuple[str, str]:
-    """
-    Read the config at the given fs_url.
-    If no fs_url is given, try the default locations.
-    """
-    _, fs_url = config_path(fs_url)
-    dirname, filename = path_split(fs_url)
-    with fs.open_fs(dirname) as confdir:
-        return (fs_url, confdir.readtext(filename))
-
-
-class NaturalOrderGroup(click.Group):
-    def list_commands(self, ctx):
-        return self.commands.keys()
-
-
-class TagType(click.ParamType):
-    name = "tag"
-
-    def convert(self, value, param, ctx):
-        if not value:
-            return tuple()
-        return tuple(tag.strip() for tag in value.split(","))
-
-
-CLI_CONFIG = click.argument(
-    "config",
-    required=False,
-    type=str,
-)
-CLI_WORKING_DIR_OPTION = click.option(
-    "--working-dir",
-    default="",
-    type=str,
-    help="The working directory",
-)
-CLI_TAGS = click.option(
-    "--tags",
-    type=TagType(),
-    default="",
-    help="tags to run",
-)
-CLI_SKIP_TAGS = click.option(
-    "--skip-tags",
-    type=TagType(),
-    default="",
-    help="tags to skip",
-)
+    webbrowser.open(uri)
 
 
 def execute(
     config: Optional[str],
-    working_dir: str,
+    working_dir: Optional[Path],
+    format: OutputFormat,
+    tags: Tags,
+    skip_tags: Tags,
     simulate: bool,
-    tags: Optional[Tuple[str]] = None,
-    skip_tags: Optional[Tuple[str]] = None,
-):
-    from schema import SchemaError
+) -> None:
+    output = JSONL() if format == "jsonl" else Default()
+    config_path = find_config(name_or_path=config)
+    Config.from_path(config_path).execute(
+        simulate=simulate,
+        output=output,
+        tags=tags,
+        skip_tags=skip_tags,
+        working_dir=working_dir or Path("."),
+    )
 
-    from . import core
 
-    config_path, config_text = read_config(config)
-
+def new(config: Optional[str]) -> None:
     try:
-        console.info(config=config_path, working_dir=working_dir)
-        core.run(
-            rules=config_text,
-            simulate=simulate,
-            working_dir=working_dir,
-            tags=tags,
-            skip_tags=skip_tags,
+        config_path = find_config(config)
+        console.print(
+            f'Config "{config_path}" already exists.\n'
+            r'Use "organize new \[name]" to create a config in the default location.'
         )
-    except NeedsMigrationError as e:
-        from .migration import MIGRATION_DOCS_URL
-
-        console.error(e, title="Config needs migration")
-        console.warn(
-            "Your config file needs some updates to work with organize v2.\n"
-            "Please see the migration guide at\n\n"
-            "%s" % MIGRATION_DOCS_URL
-        )
-        sys.exit(1)
-    except SchemaError as e:
-        console.error("Invalid config file!")
-        for err in e.autos:
-            if err and len(err) < 200:
-                core.highlighted_console.print(err)
-    except Exception as e:
-        core.highlighted_console.print_exception()
-    except (EOFError, KeyboardInterrupt):
-        console.status.stop()
-        console.warn("Aborted")
+    except ConfigNotFound as e:
+        assert e.init_path is not None
+        e.init_path.write_text(EXAMPLE_CONFIG)
+        console.print(f'Config "{e.init_path.stem}" created at "{e.init_path}"')
 
 
-@click.group(
-    help=__doc__,
-    cls=NaturalOrderGroup,
-    context_settings=dict(help_option_names=["-h", "--help"]),
-)
-@click.version_option(__version__)
-def cli():
-    pass
-
-
-@cli.command()
-@CLI_CONFIG
-@CLI_WORKING_DIR_OPTION
-@CLI_TAGS
-@CLI_SKIP_TAGS
-def run(config: Optional[str], working_dir: str, tags, skip_tags):
-    """Organizes your files according to your rules."""
-    execute(
-        config=config,
-        working_dir=working_dir,
-        simulate=False,
-        tags=tags,
-        skip_tags=skip_tags,
-    )
-
-
-@cli.command()
-@CLI_CONFIG
-@CLI_WORKING_DIR_OPTION
-@CLI_TAGS
-@CLI_SKIP_TAGS
-def sim(config: Optional[str], working_dir: str, tags, skip_tags):
-    """Simulates a run (does not touch your files)."""
-    execute(
-        config=config,
-        working_dir=working_dir,
-        simulate=True,
-        tags=tags,
-        skip_tags=skip_tags,
-    )
-
-
-@cli.command()
-@click.argument("config", required=False, type=str)
-@click.option(
-    "--editor",
-    envvar="EDITOR",
-    help="The editor to use. (Default: $EDITOR)",
-)
-def edit(config: Optional[str], editor):
-    """Edit the rules.
-
-    If called without arguments it will open the default config file in $EDITOR.
-    """
-    is_syspath, confpath = config_path(config)
-    if is_syspath:
-        click.edit(filename=confpath, editor=editor)
+def edit(config: Optional[str]) -> None:
+    config_path = find_config(config)
+    editor = os.getenv("EDITOR")
+    if editor:
+        os.system(f'{editor} "{config_path}"')
     else:
-        click.echo(f"Not a local config path: {confpath}")
+        _open_uri(config_path.as_uri())
 
 
-@cli.command()
-@CLI_CONFIG
-@click.option("--debug", is_flag=True, help="Verbose output")
-def check(config: str, debug):
-    """Checks whether a given config file is valid.
-
-    If called without arguments it will check the default config file.
-    """
-    from . import migration
-    from .config import cleanup, load_from_string, validate
-    from .core import highlighted_console as out
-    from .core import replace_with_instances
-
-    try:
-        config_path, config_str = read_config(config)
-        print(f"Checking: {config_path}")
-
-        if debug:
-            out.rule("Raw", align="left")
-            out.print(config_str)
-
-        rules = load_from_string(config_str)
-
-        if debug:
-            out.print("\n\n")
-            out.rule("Loaded", align="left")
-            out.print(rules)
-
-        rules = cleanup(rules)
-
-        if debug:
-            out.print("\n\n")
-            out.rule("Cleaned", align="left")
-            out.print(rules)
-
-        if debug:
-            out.print("\n\n")
-            out.rule("Migration from v1", align="left")
-
-        migration.migrate_v1(rules)
-
-        if debug:
-            out.print("Not needed.")
-            out.print("\n\n")
-            out.rule("Schema validation", align="left")
-
-        validate(rules)
-
-        if debug:
-            out.print("Validation ok.")
-            out.print("\n\n")
-            out.rule("Instantiation", align="left")
-
-        # warnings = replace_with_instances(rules)
-        # if debug:
-        #     out.print(rules)
-        #     for msg in warnings:
-        #         out.print("Warning: %s" % msg)
-
-        if debug:
-            out.print("\n\n")
-            out.rule("Result", align="left")
-        out.print("Config is valid.")
-
-    except Exception as e:
-        out.print_exception()
-        sys.exit(1)
+def check(config: Optional[str]) -> None:
+    config_path = find_config(config)
+    Config.from_path(config_path=config_path)
+    console.print(f'No problems found in "{config_path}".')
 
 
-@cli.command()
-@CLI_CONFIG
-@click.option("--path", is_flag=True, help="Print the path instead of revealing it.")
-def reveal(config: Optional[str], path: bool):
-    """Reveals the default config file."""
-    is_syspath, confpath = config_path(config)
-    if path:
-        click.echo(confpath)
-        return
-    try:
-        # convert the url
-        dirname, _ = path_split(confpath)
-        import webbrowser
+def debug(config: Optional[str]) -> None:
+    from rich.pretty import pprint
 
-        with fs.open_fs(dirname) as dirfs:
-            dir_url = dirfs.geturl("/")
-            if not is_syspath:
-                raise ValueError("not a local path")
-            webbrowser.open(dir_url)
-    except Exception as e:
-        click.echo(f"Cannot reveal this config ({e})")
-        click.echo(confpath)
-
-
-@cli.command()
-def schema():
-    """Prints the json schema for config files."""
-    # ORGANIZE SCHEMA IS DEPRECATED AND WILL BE REMOVED IN THE FUTURE
-    import json
-
-    from .config import CONFIG_SCHEMA
-    from .console import console as richconsole
-
-    js = json.dumps(
-        CONFIG_SCHEMA.json_schema(
-            schema_id="https://tfeldmann.de/organize.schema.json",
-        )
+    config_path = find_config(config)
+    pprint(
+        Config.from_path(config_path=config_path),
+        expand_all=True,
+        indent_guides=False,
     )
-    richconsole.print_json(js)
 
 
-@cli.command()
-def docs():
-    """Opens the documentation."""
-    import webbrowser
+def show(config: Optional[str], path: bool, reveal: bool) -> None:
+    config_path = find_config(name_or_path=config)
+    if path:
+        print(config_path)
+    elif reveal:
+        _open_uri(config_path.parent.as_uri())
+    else:
+        syntax = Syntax(config_path.read_text(), "yaml")
+        console.print(syntax)
 
-    webbrowser.open(DOCS_RTD)
+
+def list_() -> None:
+    table = Table()
+    table.add_column("Config")
+    table.add_column("Path", no_wrap=True, style="dim")
+    for path in list_configs():
+        table.add_row(path.stem, str(path))
+    console.print(table)
+
+
+def docs() -> None:
+    uri = DOCS_GHPAGES if __is_prerelease__ else DOCS_RTD
+    print(f'Opening "{uri}"')
+    _open_uri(uri=uri)
+
+
+class CliArgs(BaseModel):
+    model_config = ConfigDict(extra="forbid")
+
+    # commands
+    run: bool
+    sim: bool
+    new: bool
+    edit: bool
+    check: bool
+    debug: bool
+    show: bool
+    list: bool
+    docs: bool
+
+    # run / sim options
+    config: Optional[str] = Field(..., alias="<config>")
+    working_dir: Optional[Path] = Field(..., alias="--working-dir")
+    format: OutputFormat = Field("default", alias="--format")
+    tags: Optional[str] = Field(..., alias="--tags")
+    skip_tags: Optional[str] = Field(..., alias="--skip-tags")
+
+    # show options
+    path: bool = Field(False, alias="--path")
+    reveal: bool = Field(False, alias="--reveal")
+
+    # docopt options
+    version: bool = Field(..., alias="--version")
+    help: bool = Field(..., alias="--help")
+
+    @field_validator("tags", "skip_tags", mode="after")
+    @classmethod
+    def split_tags(cls, val) -> Set[str]:
+        if val is None:
+            return set()
+        return set(val.split(","))
+
+
+def cli() -> None:
+    arguments = docopt(
+        __doc__,
+        version=f"organize v{__version__}",
+        default_help=True,
+    )
+    try:
+        args = CliArgs.model_validate(arguments)
+        _execute = partial(
+            execute,
+            config=args.config,
+            working_dir=args.working_dir,
+            format=args.format,
+            tags=args.tags,
+            skip_tags=args.skip_tags,
+        )
+        if args.run:
+            _execute(simulate=False)
+        elif args.sim:
+            _execute(simulate=True)
+        elif args.new:
+            new(config=args.config)
+        elif args.edit:
+            edit(config=args.config)
+        elif args.check:
+            check(config=args.config)
+        elif args.debug:
+            debug(config=args.config)
+        elif args.show:
+            show(config=args.config, path=args.path, reveal=args.reveal)
+        elif args.list:
+            list_()
+        elif args.docs:
+            docs()
+    except (ConfigError, ConfigNotFound) as e:
+        console.print(f"[red]Error: Config problem[/]\n{e}")
+        sys.exit(1)
+    except ValidationError as e:
+        console.print(f"[red]Error: Invalid CLI arguments[/]\n{e}")
+        sys.exit(2)
+    except ScannerError as e:
+        console.print(f"[red]Error: YAML syntax error[/]\n{e}")
+        sys.exit(3)
 
 
 if __name__ == "__main__":
