@@ -1,1 +1,2 @@
-__version__ = "2.4.3"
+__version__ = "3.0.0a0"
+__is_prerelease__ = True
