@@ -1,6 +1,3 @@
-from typing import Dict, Type
-
-from .action import Action
 from .confirm import Confirm
 from .copy import Copy
 from .delete import Delete
@@ -14,17 +11,17 @@ from .symlink import Symlink
 from .trash import Trash
 from .write import Write
 
-ACTIONS = {
-    Confirm.name: Confirm,
-    Copy.name: Copy,
-    Delete.name: Delete,
-    Echo.name: Echo,
-    MacOSTags.name: MacOSTags,
-    Move.name: Move,
-    Python.name: Python,
-    Rename.name: Rename,
-    Shell.name: Shell,
-    Symlink.name: Symlink,
-    Trash.name: Trash,
-    Write.name: Write,
-}  # type: Dict[str, Type[Action]]
+ALL = (
+    Confirm,
+    Copy,
+    Delete,
+    Echo,
+    MacOSTags,
+    Move,
+    Python,
+    Rename,
+    Shell,
+    Symlink,
+    Trash,
+    Write,
+)
