@@ -1,21 +1,23 @@
-import logging
 import sys
+from typing import ClassVar
 
 import simplematch as sm
-from schema import Or
+from pydantic.config import ConfigDict
+from pydantic.dataclasses import dataclass
 
-from organize.utils import Template
+from organize.action import ActionConfig
+from organize.output import Output
+from organize.resource import Resource
+from organize.template import Template, render
+from organize.validators import FlatList
 
-from .action import Action
 
-logger = logging.getLogger(__name__)
-
-
-class MacOSTags(Action):
+@dataclass(config=ConfigDict(coerce_numbers_to_str=True, extra="forbid"))
+class MacOSTags:
 
     """Add macOS tags.
 
-    Args:
+    Attributes:
         *tags (str): A list of tags or a single tag.
 
     The color can be specified in brackets after the tag name, for example:
@@ -28,43 +30,45 @@ class MacOSTags(Action):
     `orange`.
     """
 
-    name = "macos_tags"
+    tags: FlatList[str]
 
-    @classmethod
-    def get_schema(cls):
-        return {cls.name: Or(str, [str])}
+    action_config: ClassVar[ActionConfig] = ActionConfig(
+        name="macos_tags",
+        standalone=False,
+        files=True,
+        dirs=True,
+    )
 
-    def __init__(self, *tags):
-        self.tags = [Template.from_string(tag) for tag in tags]
-
-    def pipeline(self, args: dict, simulate: bool):
-        fs = args["fs"]
-        fs_path = args["fs_path"]
-        path = fs.getsyspath(fs_path)
-
+    def __post_init__(self):
+        self._tags = [Template.from_string(tag) for tag in self.tags]
         if sys.platform != "darwin":
             raise EnvironmentError("The macos_tags action is only available on macOS")
 
+    def pipeline(self, res: Resource, output: Output, simulate: bool):
         import macos_tags
 
         COLORS = [c.name.lower() for c in macos_tags.Color]
 
-        for template in self.tags:
-            tag = template.render(**args)
+        for template in self._tags:
+            tag = render(template, res.dict())
             name, color = self._parse_tag(tag)
 
             if color not in COLORS:
                 raise ValueError(
-                    f"color {color} is unknown. (Available: {' / '.join(COLORS)})"
+                    "color %s is unknown. (Available: %s)" % (color, " / ".join(COLORS))
                 )
 
-            self.print(f'Adding tag: "{name}" (color: {color})')
+            output.msg(
+                res=res,
+                sender=self,
+                msg=f'Adding tag: "{name}" (color: {color})',
+            )
             if not simulate:
                 _tag = macos_tags.Tag(
                     name=name,
                     color=macos_tags.Color[color.upper()],
                 )  # type: ignore
-                macos_tags.add(_tag, file=str(path))
+                macos_tags.add(_tag, file=str(res.path))
 
     def _parse_tag(self, s):
         """parse a tag definition and return a tuple (name, color)"""
@@ -72,6 +76,3 @@ class MacOSTags(Action):
         if not result:
             return s, "none"
         return result["name"], result["color"].lower()
-
-    def __str__(self) -> str:
-        return f"MacOSTags(tags={self.tags})"
