@@ -1,29 +1,24 @@
-import logging
-import os
-from typing import Callable
+from __future__ import annotations
 
-from fs import path
-from fs.base import FS
-from fs.move import move_dir, move_file
-from schema import Optional, Or
+from typing import ClassVar
 
-from organize.utils import Template, safe_description
+from pydantic.config import ConfigDict
+from pydantic.dataclasses import dataclass
 
-from ._conflict_resolution import (
-    CONFLICT_OPTIONS,
-    check_conflict,
-    resolve_overwrite_conflict,
-)
-from .action import Action
+from organize.action import ActionConfig
+from organize.output import Output
+from organize.resource import Resource
+from organize.template import Template, render
 
-logger = logging.getLogger(__name__)
+from .common.conflict import ConflictMode, resolve_conflict
 
 
-class Rename(Action):
+@dataclass(config=ConfigDict(coerce_numbers_to_str=True, extra="forbid"))
+class Rename:
 
     """Renames a file.
 
-    Args:
+    Attributes:
         name (str):
             The new name for the file / dir.
 
@@ -39,75 +34,45 @@ class Rename(Action):
     The next action will work with the renamed file / dir.
     """
 
-    name = "rename"
-    arg_schema = Or(
-        str,
-        {
-            "name": str,
-            Optional("on_conflict"): Or(*CONFLICT_OPTIONS),
-            Optional("rename_template"): str,
-        },
+    new_name: str
+    on_conflict: ConflictMode = "rename_new"
+    rename_template: str = "{name} {counter}{extension}"
+    # TODO: keep_extension?
+
+    action_config: ClassVar[ActionConfig] = ActionConfig(
+        name="rename",
+        standalone=False,
+        files=True,
+        dirs=True,
     )
 
-    def __init__(
-        self,
-        name: str,
-        on_conflict="rename_new",
-        rename_template="{name} {counter}{extension}",
-    ) -> None:
-        if on_conflict not in CONFLICT_OPTIONS:
-            raise ValueError(
-                f"on_conflict must be one of {', '.join(CONFLICT_OPTIONS)}"
-            )
+    def __post_init__(self):
+        self._new_name = Template.from_string(self.new_name)
+        self._rename_template = Template.from_string(self.rename_template)
 
-        self.new_name = Template.from_string(name)
-        self.conflict_mode = on_conflict
-        self.rename_template = Template.from_string(rename_template)
-
-    def pipeline(self, args: dict, simulate: bool):
-        fs = args["fs"]  # type: FS
-        src_path = args["fs_path"]
-
-        new_name = self.new_name.render(**args)
+    def pipeline(self, res: Resource, output: Output, simulate: bool):
+        assert res.path is not None, "Does not support standalone mode"
+        new_name = render(self._new_name, res.dict())
         if "/" in new_name:
             raise ValueError(
                 "The new name cannot contain slashes. "
                 "To move files or folders use `move`."
             )
+        dst = res.path.with_name(new_name)
+        skip_action, dst = resolve_conflict(
+            dst=dst,
+            res=res,
+            conflict_mode=self.on_conflict,
+            rename_template=self._rename_template,
+            simulate=simulate,
+            output=output,
+        )
 
-        dst_path = path.join(path.dirname(src_path), new_name)
+        if skip_action:
+            return
 
-        if dst_path == src_path:
-            self.print("Name did not change")
-        else:
-            move_action: Callable[[FS, str, FS, str], None]
-            if fs.isdir(src_path):
-                move_action = move_dir
-            elif fs.isfile(src_path):
-                move_action = move_file
-
-            # check for conflicts
-            skip, dst_path = check_conflict(
-                src_fs=fs,
-                src_path=src_path,
-                dst_fs=fs,
-                dst_path=dst_path,
-                conflict_mode=self.conflict_mode,
-                rename_template=self.rename_template,
-                simulate=simulate,
-                print=self.print,
-            )
-
-            if not skip:
-                self.print(f"Rename to {safe_description(fs, dst_path)}")
-                if not simulate:
-                    move_action(fs, src_path, fs, dst_path)
-
-        # the next action should work with the renamed file
-        return {
-            "fs": fs,
-            "fs_path": "./" + dst_path,
-        }
-
-    def __str__(self) -> str:
-        return f"Rename(new_name={self.new_name}, conflict_mode={self.conflict_mode})"
+        output.msg(res=res, msg=f"Renaming to {new_name}", sender=self)
+        if not simulate:
+            res.path.rename(dst)
+        res.path = dst
+        res.walker_skip_pathes.add(dst)
