@@ -1,20 +1,23 @@
-import logging
 import subprocess
-from subprocess import PIPE
+from typing import ClassVar
 
-from schema import Optional, Or
+from pydantic.config import ConfigDict
+from pydantic.dataclasses import dataclass
 
-from ..utils import Template
-from .action import Action
+from organize.action import ActionConfig
+from organize.output import Output
+from organize.resource import Resource
+from organize.template import Template, render
 
-logger = logging.getLogger(__name__)
+# TODO: Terminal waterfall: https://github.com/Textualize/rich/discussions/2985
 
 
-class Shell(Action):
+@dataclass(config=ConfigDict(coerce_numbers_to_str=True, extra="forbid"))
+class Shell:
     """
     Executes a shell command
 
-    Args:
+    Attributes:
         cmd (str): The command to execute.
         run_in_simulation (bool):
             Whether to execute in simulation mode (default = false)
@@ -31,70 +34,51 @@ class Shell(Action):
     - `{shell.returncode}` (`int`): The returncode of the executed process.
     """
 
-    name = "shell"
-    arg_schema = Or(
-        str,
-        {
-            "cmd": str,
-            Optional("run_in_simulation"): bool,
-            Optional("ignore_errors"): bool,
-            Optional("simulation_output"): str,
-            Optional("simulation_returncode"): int,
-        },
+    cmd: str
+    run_in_simulation: bool = False
+    ignore_errors: bool = False
+    simulation_output: str = "** simulation **"
+    simulation_returncode: int = 0
+
+    action_config: ClassVar[ActionConfig] = ActionConfig(
+        name="shell",
+        standalone=True,
+        files=True,
+        dirs=True,
     )
 
-    def __init__(
-        self,
-        cmd: str,
-        run_in_simulation=False,
-        ignore_errors=False,
-        simulation_output="** simulation **",
-        simulation_returncode=0,
-    ):
-        self.cmd = Template.from_string(cmd)
-        self.run_in_simulation = run_in_simulation
-        self.ignore_errors = ignore_errors
-        self.simulation_output = Template.from_string(simulation_output)
-        self.simulation_returncode = simulation_returncode
+    def __post_init__(self):
+        self._cmd = Template.from_string(self.cmd)
+        self._simulation_output = Template.from_string(self.simulation_output)
 
-    def pipeline(self, args: dict, simulate: bool):
-        full_cmd = self.cmd.render(**args)
+    def pipeline(self, res: Resource, output: Output, simulate: bool):
+        full_cmd = render(self._cmd, res.dict())
 
         if not simulate or self.run_in_simulation:
-            # we use call instead of run to be compatible with python < 3.5
-            self.print(f"$ {full_cmd}")
-            logger.info('Executing command "%s" in shell.', full_cmd)
+            output.msg(res=res, msg=f"$ {full_cmd}", sender=self)
             try:
                 call = subprocess.run(
                     full_cmd,
                     check=True,
-                    stdout=PIPE,
+                    stdout=subprocess.PIPE,
                     stderr=subprocess.STDOUT,
                     shell=True,
                 )
-                return {
-                    self.get_name(): {
-                        "output": call.stdout.decode("utf-8"),
-                        "returncode": 0,
-                    }
-                }
             except subprocess.CalledProcessError as e:
                 if not self.ignore_errors:
                     raise e
-                return {
-                    self.get_name(): {
-                        "output": e.stdout.decode("utf-8"),
-                        "returncode": e.returncode,
-                    }
-                }
-        else:
-            self.print(f"** not run in simulation ** $ {full_cmd}")
-            return {
-                self.get_name(): {
-                    "output": self.simulation_output.render(**args),
-                    "returncode": self.simulation_returncode,
-                }
-            }
 
-    def __str__(self) -> str:
-        return f'Shell(cmd="{self.cmd}")'
+            res.vars[self.action_config.name] = {
+                "output": call.stdout.decode("utf-8"),
+                "returncode": call.returncode,
+            }
+        else:
+            output.msg(
+                res=res,
+                msg=f"** not run in simulation ** $ {full_cmd}",
+                sender=self,
+            )
+            res.vars[self.action_config.name] = {
+                "output": render(self._simulation_output, res.dict()),
+                "returncode": self.simulation_returncode,
+            }
