@@ -1,28 +1,23 @@
-from fs.base import FS
+from typing import ClassVar
 
-from .filter import Filter, FilterResult
+from pydantic.config import ConfigDict
+from pydantic.dataclasses import dataclass
+
+from organize.filter import FilterConfig
+from organize.output import Output
+from organize.resource import Resource
 
 
-class Empty(Filter):
+@dataclass(config=ConfigDict(extra="forbid"))
+class Empty:
 
     """Finds empty dirs and files"""
 
-    name = "empty"
+    filter_config: ClassVar[FilterConfig] = FilterConfig(
+        name="empty",
+        files=True,
+        dirs=True,
+    )
 
-    @classmethod
-    def get_schema(cls):
-        return cls.get_name_schema()
-
-    def pipeline(self, args: dict) -> FilterResult:
-        fs = args["fs"]  # type: FS
-        fs_path = args["fs_path"]  # type: str
-
-        if fs.isdir(fs_path):
-            result = fs.isempty(fs_path)
-        else:
-            result = fs.getsize(fs_path) == 0
-
-        return FilterResult(matches=result, updates={})
-
-    def __str__(self) -> str:
-        return "Empty()"
+    def pipeline(self, res: Resource, output: Output) -> bool:
+        return res.is_empty()
