@@ -1,15 +1,16 @@
 import operator
 import re
-from typing import Callable, Dict
-from typing import Optional as Opt
-from typing import Sequence, Set, Tuple
+from pathlib import Path
+from typing import ClassVar, Iterable
 
-from fs.filesize import binary, decimal, traditional
-from schema import Optional, Or
+from pydantic import Field
+from pydantic.config import ConfigDict
+from pydantic.dataclasses import dataclass
 
-from organize.utils import flattened_string_list
-
-from .filter import Filter, FilterResult
+from organize.filter import FilterConfig
+from organize.output import Output
+from organize.resource import Resource
+from organize.validators import FlatList
 
 OPERATORS = {
     "<": operator.lt,
@@ -20,12 +21,30 @@ OPERATORS = {
     ">=": operator.ge,
     ">": operator.gt,
 }
+
 SIZE_REGEX = re.compile(
     r"^(?P<op>[<>=]*)(?P<num>(\d*\.)?\d+)(?P<unit>[kmgtpezy]?i?)b?$"
 )
 
 
-def create_constraints(inp: str) -> Set[Tuple[Callable[[int, int], bool], int]]:
+def read_file_size(path: Path) -> int:
+    return path.stat().st_size
+
+
+def read_dir_size(path: Path) -> int:
+    return sum(f.stat().st_size for f in path.glob("**/*") if f.is_file())
+
+
+def read_resource_size(res: Resource) -> int:
+    assert res.path is not None
+    if res.is_file():
+        return read_file_size(res.path)
+    if res.is_dir():
+        return read_dir_size(res.path)
+    raise ValueError("Unknown file type")
+
+
+def create_constraints(inp: str):
     """
     Given an input string it returns a list of tuples (comparison operator,
     number of bytes).
@@ -34,7 +53,6 @@ def create_constraints(inp: str) -> Set[Tuple[Callable[[int, int], bool], int]]:
     Calculation is in bytes, even if the "b" is lowercase. If an "i" is present
     we calculate base 1024.
     """
-    result = set()  # type: Set[Tuple[Callable[[int, int], bool], int]]
     parts = str(inp).replace(" ", "").lower().split(",")
     for part in parts:
         try:
@@ -47,20 +65,55 @@ def create_constraints(inp: str) -> Set[Tuple[Callable[[int, int], bool], int]]:
                 base = 1024 if unit.endswith("i") else 1000
                 exp = "kmgtpezy".index(unit[0]) + 1 if unit else 0
                 numbytes = num * base**exp
-                result.add((op, numbytes))
+                yield (op, numbytes)
         except (AttributeError, KeyError, IndexError, ValueError, TypeError) as e:
-            raise ValueError(f"Invalid size format: {part}") from e
-    return result
+            raise ValueError("Invalid size format: %s" % part) from e
 
 
 def satisfies_constraints(size, constraints):
     return all(op(size, p_size) for op, p_size in constraints)
 
 
-class Size(Filter):
+def number_with_unit(size: int, suffixes: Iterable[str], base: int) -> str:
+    size = int(size)
+    if size == 1:
+        return "1 byte"
+    elif size < base:
+        return "{:,} bytes".format(size)
+
+    for i, suffix in enumerate(suffixes, 2):
+        unit = base**i
+        if size < unit:
+            break
+    return "{:,.1f} {}".format((base * size / unit), suffix)
+
+
+def traditional(size):
+    """Convert a filesize in to a string (powers of 1024, JDEC prefixes)."""
+    return number_with_unit(
+        size, ("KB", "MB", "GB", "TB", "PB", "EB", "ZB", "YB"), 1024
+    )
+
+
+def binary(size):
+    """Convert a filesize in to a string (powers of 1024, IEC prefixes)."""
+    return number_with_unit(
+        size, ("KiB", "MiB", "GiB", "TiB", "PiB", "EiB", "ZiB", "YiB"), 1024
+    )
+
+
+def decimal(size):
+    """Convert a filesize in to a string (powers of 1000, SI prefixes)."""
+    return number_with_unit(
+        size, ("kB", "MB", "GB", "TB", "PB", "EB", "ZB", "YB"), 1000
+    )
+
+
+@dataclass(config=ConfigDict(coerce_numbers_to_str=True, extra="forbid"))
+class Size:
     """Matches files and folders by size
 
-    Args:
+    Attributes:
         *conditions (list(str) or str):
             The size constraints.
 
@@ -83,42 +136,30 @@ class Size(Filter):
     - `{size.decimal}`: (str) Size with unit (powers of 1000, SI prefixes)
     """
 
-    name = "size"
-    arg_schema = Or(object, [object])
-    schema_support_instance_without_args = True
+    conditions: FlatList[str] = Field(default_factory=list)
 
-    def __init__(self, *conditions: Sequence[str]) -> None:
-        self.conditions = ", ".join(flattened_string_list(list(conditions)))
-        self.constraints = create_constraints(self.conditions)
+    filter_config: ClassVar[FilterConfig] = FilterConfig(
+        name="size", files=True, dirs=True
+    )
+
+    def __post_init__(self):
+        self._constraints = set()
+        for x in self.conditions:
+            for constraint in create_constraints(x):
+                self._constraints.add(constraint)
 
     def matches(self, filesize: int) -> bool:
-        if not self.constraints:
+        if not self._constraints:
             return True
-        return all(op(filesize, c_size) for op, c_size in self.constraints)
+        return all(op(filesize, c_size) for op, c_size in self._constraints)
 
-    def pipeline(self, args: dict) -> FilterResult:
-        fs = args["fs"]
-        fs_path = args["fs_path"]
-
-        if fs.isdir(fs_path):
-            size = sum(
-                info.size
-                for _, info in fs.walk.info(path=fs_path, namespaces=["details"])
-            )
-        else:
-            size = fs.getsize(fs_path)
-
-        return FilterResult(
-            matches=self.matches(size),
-            updates={
-                self.get_name(): {
-                    "bytes": size,
-                    "traditional": traditional(size),
-                    "binary": binary(size),
-                    "decimal": decimal(size),
-                },
-            },
-        )
-
-    def __str__(self) -> str:
-        return f"FileSize({' '.join(self.conditions)})"
+    def pipeline(self, res: Resource, output: Output) -> bool:
+        assert res.path is not None
+        bytes = read_resource_size(res=res)
+        res.vars[self.filter_config.name] = {
+            "bytes": bytes,
+            "traditional": traditional(bytes),
+            "binary": binary(bytes),
+            "decimal": decimal(bytes),
+        }
+        return self.matches(bytes)
