@@ -1,25 +1,41 @@
+from collections import Counter
 from datetime import datetime
-from unittest.mock import patch
 
-from organize.actions import Echo
+from conftest import make_files
 
-
-def test_echo_basic():
-    echo = Echo("Hello World")
-    with patch.object(echo, "print") as m:
-        echo.run(simulate=False)
-        m.assert_called_with("Hello World")
+from organize import Config
 
 
-def test_echo_args():
-    echo = Echo('Date formatting: {now.strftime("%Y-%m-%d")}')
-    with patch.object(echo, "print") as m:
-        echo.run(simulate=False, now=datetime(2019, 1, 5))
-        m.assert_called_with("Date formatting: 2019-01-05")
+def test_echo_basic(testoutput):
+    Config.from_string(
+        """
+        rules:
+          - actions:
+              - echo: "Hello World"
+        """
+    ).execute(simulate=False, output=testoutput)
+    assert testoutput.messages == ["Hello World"]
 
 
-def test_echo_path():
-    echo = Echo("{year}")
-    with patch.object(echo, "print") as m:
-        echo.run(simulate=False, year=2017)
-        m.assert_called_with("2017")
+def test_echo_args(testoutput):
+    Config.from_string(
+        """
+        rules:
+          - actions:
+              - echo: "Date formatting: {now().strftime('%Y')}"
+        """
+    ).execute(simulate=False, output=testoutput)
+    assert testoutput.messages == [f"Date formatting: {datetime.now().year}"]
+
+
+def test_echo_path(fs, testoutput):
+    make_files(["fileA.txt", "fileB.txt"], "test")
+    Config.from_string(
+        """
+        rules:
+          - locations: /test
+            actions:
+              - echo: "{path.stem}"
+        """
+    ).execute(simulate=False, output=testoutput)
+    assert Counter(testoutput.messages) == Counter(["fileA", "fileB"])
