@@ -1,59 +1,87 @@
 import pytest
 from conftest import make_files, read_files
-from fs.base import FS
 
-from organize import core
-
-files = {
-    "test.txt": "",
-    "file.txt": "Hello world\nAnother line",
-    "another.txt": "",
-    "folder": {
-        "x.txt": "",
-    },
-}
+from organize.config import Config
 
 
-@pytest.fixture
-def testfiles(testfs) -> FS:
-    make_files(testfs, files)
-    yield testfs
-
-
-def test_copy_on_itself(testfiles):
+def test_move_onto_itself(fs):
+    FILES = {
+        "test.txt": "",
+        "file.txt": "Hello world\nAnother line",
+        "another.txt": "",
+        "folder": {
+            "x.txt": "",
+        },
+    }
+    make_files(FILES, "test")
     config = """
     rules:
-      - locations: "/"
+      - locations: "test"
         actions:
-          - move: "/"
+          - move: "test"
     """
-    core.run(config, simulate=True, working_dir=testfiles)
-    result = read_files(testfiles)
-    assert result == files
+    Config.from_string(config).execute(simulate=False)
+    assert read_files("test") == FILES
 
 
 @pytest.mark.parametrize(
-    "mode,files,test_txt_content",
+    "mode,result",
     [
-        ("skip", ["file.txt", "test.txt"], "old"),
-        ("overwrite", ["test.txt"], "new"),
-        ("rename_new", ["test.txt", "test 2.txt"], "old"),
-        ("rename_existing", ["test.txt", "test 2.txt"], "new"),
+        ("skip", {"src.txt": "src", "dst.txt": "dst"}),
+        ("overwrite", {"dst.txt": "src"}),
+        ("rename_new", {"dst 2.txt": "src", "dst.txt": "dst"}),
+        ("rename_existing", {"dst.txt": "src", "dst 2.txt": "dst"}),
     ],
 )
-def test_move_conflict(testfs: FS, mode, files, test_txt_content):
+def test_move_conflict(fs, mode, result):
+    make_files(
+        {
+            "src.txt": "src",
+            "dst.txt": "dst",
+        },
+        path="test",
+    )
+    # src is moved onto dst.
     config = f"""
     rules:
-      - locations: "/"
+      - locations: "test"
         filters:
-          - name: file
+          - name: src
         actions:
           - move:
-              dest: "test.txt"
+              dest: "{{location}}/dst.txt"
               on_conflict: {mode}
     """
-    testfs.writetext("file.txt", "new")
-    testfs.writetext("test.txt", "old")
-    core.run(config, simulate=False, working_dir=testfs)
-    assert set(testfs.listdir("/")) == set(files)
-    assert testfs.readtext("test.txt") == test_txt_content
+    Config.from_string(config).execute(simulate=False)
+    assert read_files("test") == result
+
+
+def test_move_folder_conflict(fs):
+    make_files(
+        {
+            "src": {"dir": {"src.txt": ""}},
+            "dst": {"dir": {"dst.txt": ""}},
+        },
+        "test",
+    )
+    # src is moved onto dst.
+    Config.from_string(
+        """
+        rules:
+            -   locations: "/test/src"
+                targets: dirs
+                filters:
+                    - name: dir
+                actions:
+                    - move:
+                        dest: "{location}/../dst"
+                        on_conflict: "rename_new"
+        """
+    ).execute(simulate=False)
+    assert read_files("test") == {
+        "src": {},
+        "dst": {
+            "dir": {"dst.txt": ""},
+            "dir 2": {"src.txt": ""},
+        },
+    }
