@@ -1,48 +1,35 @@
 from unittest.mock import patch
 
-import fs
-
-from organize.actions import Trash
-from organize.core import run
+from organize import Config
 
 
-def test_trash_mocked():
-    with patch("send2trash.send2trash") as mck:
-        trash = Trash()
-        trash.trash(path="~/Desktop/Test.zip", simulate=False)
-        mck.assert_called_with("~/Desktop/Test.zip")
+def test_trash_mocked(tmp_path):
+    testfile = tmp_path / "test.txt"
+    testfile.touch()
+    with patch("organize.actions.trash.trash") as mck:
+        Config.from_string(
+            f"""
+            rules:
+              - locations: {tmp_path}
+                actions:
+                  - trash
+            """
+        ).execute(simulate=False)
+        mck.assert_called_once_with(testfile)
 
 
-def test_trash_file():
-    config = """
-    rules:
-      - locations: "."
-        filters:
-          - name: "test-trash-209318123"
-        actions:
-          - trash
-    """
-    with fs.open_fs("temp://") as tmp:
-        tmp.writetext("test-trash-209318123.txt", "Content")
-        assert tmp.exists("test-trash-209318123.txt")
-        run(config, simulate=False, working_dir=tmp.getsyspath("/"))
-        assert not tmp.exists("test-trash-209318123.txt")
-
-
-def test_trash_folder():
-    config = """
-    rules:
-      - locations: "."
-        targets: dirs
-        filters:
-          - name: "test-trash-209318123"
-        actions:
-          - trash
-    """
-    with fs.open_fs("temp://") as tmp:
-        tmp_dir = tmp.makedir("test-trash-209318123")
-        tmp_dir.touch("file.txt")
-        assert tmp.exists("test-trash-209318123/file.txt")
-        run(config, simulate=False, working_dir=tmp.getsyspath("/"))
-        assert not tmp.exists("test-trash-209318123/file.txt")
-        assert not tmp.exists("test-trash-209318123")
+def test_trash_folder(tmp_path):
+    testfolder = tmp_path / "test"
+    testfolder.mkdir()
+    (testfolder / "testfile.txt").touch()
+    with patch("organize.actions.trash.trash") as mck:
+        Config.from_string(
+            f"""
+            rules:
+              - locations: {tmp_path}
+                targets: dirs
+                actions:
+                  - trash
+            """
+        ).execute(simulate=False)
+        mck.assert_called_once_with(testfolder)
