@@ -1,31 +1,32 @@
 from conftest import make_files, read_files
 
-from organize import core
+from organize import Config
 
 
-def test_dependent_rules(testfs):
+def test_dependent_rules(fs):
     files = {
         "asd.txt": "",
         "newname 2.pdf": "",
         "newname.pdf": "",
         "test.pdf": "",
     }
-    make_files(testfs, files)
-    config = """
+    make_files(files, "test")
+    Config.from_string(
+        """
     rules:
-    - locations: "."
+    - locations: /test
       filters:
         - name: test
       actions:
-        - copy: newfolder/test.pdf
-    - locations: "newfolder"
+        - copy: /test/newfolder/
+    - locations: /test/newfolder
       filters:
         - name: test
       actions:
         - rename: test-found.pdf
     """
-    core.run(config, simulate=False, working_dir=testfs)
-    assert read_files(testfs) == {
+    ).execute(simulate=False)
+    assert read_files("test") == {
         "newname.pdf": "",
         "newname 2.pdf": "",
         "test.pdf": "",
