@@ -1,49 +1,23 @@
-import pathlib
-from typing import Union
-from unittest.mock import patch
+from pathlib import Path
+from typing import Dict, List, Union
 
-import fs
 import pytest
-from fs.base import FS
-from fs.path import basename, join
 
-ORGANIZE_DIR = pathlib.Path(__file__).parent.parent
+from organize.output import SavingOutput
 
-
-@pytest.fixture
-def tempfs():
-    with fs.open_fs("temp://") as tmp:
-        yield tmp
+ORGANIZE_DIR = Path(__file__).parent.parent
 
 
 @pytest.fixture
-def memfs():
-    with fs.open_fs("mem://") as mem:
-        yield mem
+def testoutput() -> SavingOutput:
+    return SavingOutput()
 
 
-@pytest.fixture(
-    params=[
-        "mem://",
-        pytest.param("temp://"),
-    ],
-)
-def testfs(request) -> FS:
-    with fs.open_fs(request.param) as tmp:
-        yield tmp
-
-
-@pytest.fixture
-def mock_echo():
-    with patch("organize.actions.Echo.print") as mck:
-        yield mck
-
-
-def make_files(fs: FS, layout: Union[dict, list], path="/"):
+def make_files(structure: Union[Dict, List], path: Union[Path, str] = "."):
     """
-    Example layout:
+    Example structure:
 
-        layout = {
+        {
             "folder": {
                 "subfolder": {
                     "test.txt": "",
@@ -53,33 +27,43 @@ def make_files(fs: FS, layout: Union[dict, list], path="/"):
             "file.txt": "Hello world\nAnother line",
         }
     """
-    fs.makedirs(path, recreate=True)
-    if isinstance(layout, list):
-        for f in layout:
-            fs.touch(f)
+    if isinstance(path, str):
+        path = Path(path)
+    path.mkdir(parents=True, exist_ok=True)
+
+    # structure is a list of filenames
+    if isinstance(structure, list):
+        for name in structure:
+            (path / name).touch()
         return
-    for k, v in layout.items():
-        respath = join(path, k)
+
+    # structure is a dict
+    for name, content in structure.items():
+        resource: Path = path / name
 
         # folders are dicts
-        if isinstance(v, dict):
-            make_files(fs=fs, layout=v, path=respath)
+        if isinstance(content, dict):
+            make_files(structure=content, path=resource)
 
         # everything else is a file
-        elif v is None:
-            fs.touch(respath)
-        elif isinstance(v, bytes):
-            fs.writebytes(respath, v)
-        elif isinstance(v, str):
-            fs.writetext(respath, v)
+        elif content is None:
+            resource.touch()
+        elif isinstance(content, bytes):
+            resource.write_bytes(content)
+        elif isinstance(content, str):
+            resource.write_text(content)
         else:
-            raise ValueError(f"Unknown file data {v}")
+            raise ValueError(f"Unknown file data {content}")
 
 
-def read_files(fs: FS, path="/"):
+def read_files(path: Union[Path, str] = "."):
+    if isinstance(path, str):
+        path = Path(path)
+
     result = dict()
-    for x in fs.walk.files(path, max_depth=0):
-        result[basename(x)] = fs.readtext(x)
-    for x in fs.walk.dirs(path, max_depth=0):
-        result[basename(x)] = read_files(fs, path=join(path, x))
+    for x in path.glob("*"):
+        if x.is_file():
+            result[x.name] = x.read_text()
+        if x.is_dir():
+            result[x.name] = read_files(x)
     return result
